MODIFIED: MODIFIED: Seller and Store Delete function

See merge request disystem/bravomart-api!313

+++ app/Http/Controllers/Api/V1/Admin/AdminStoreManageController.php
@@ -106,6 +106,11 @@ class AdminStoreManageController extends Controller
 
     public function destroy($id)
     {
+        if (runningOrderExists($id)) {
+            return response()->json([
+                'message' => __('messages.has_running_orders', ['name' => 'Store'])
+            ]);
+        }
         $this->storeRepo->delete($id);
         return $this->success(translate('messages.delete_success'));
     }
